Slider functionality
Added slider functionality for password length
