+++ css/styles.css
@@ -60,54 +60,74 @@ button{
     border-bottom: 3px solid #9F87AF;
 }
 
-.slider-container{
+.slider-container {
     position: relative;
     width: 100%;
     margin-block: 1em;
 }
 
-
-.slider {
+.slider-value-container {
+    position: relative;
     width: 100%;
-    -webkit-appearance: none;
-    appearance: none;
-    height: 8px;
-    background: #ddd;
-    outline: none;
-    border-radius: 5px;
 }
 
 .slider-value {
-    display: none;
     position: absolute;
-    top: 0px;
-    background-color: #FFD400;
-    color: #541388;
-    padding: 5px;
-    border-radius: 50%;
-    font-size: 14px;
-    width: 30px;
     text-align: center;
+    font-size: .9rem;
+    color: #541388;
+    font-weight: 600;
+    height: 30px;
+    width: 30px;
+    bottom: -75px;
+    left: 3%;
+    transform: translateX(-50%);
+    line-height: 30px;
+    z-index: 2;
+    opacity: 0;
+    transition: opacity 0.2s ease;
+}
+
+.opaque
+{
+    opacity: 1;
+}
+
+.slider-value::after {
+    position: absolute;
+    content: '';
+    height: 30px;
+    width: 30px;
+    left: 50%;
+    background-color: #FFD400;
+    transform: translateX(-50%) rotate(45deg);
+    outline: 3px solid #E9F1F7;
+    border-bottom-left-radius: 50%;
+    border-bottom-right-radius: 50%;
+    border-top-right-radius: 50%;
+    border-top-left-radius: .15em;
+    z-index: -1;
+}
+
+.slider {
+    width: 100%;
+    appearance: none;
+    height: .5em;
+    background-color: #9F87AF;
+    border-radius: 2em;
+    outline: none;
+    border: none;
+    flex: 1;
+    cursor: pointer;
 }
 
 .slider::-webkit-slider-thumb {
-    -webkit-appearance: none;
     appearance: none;
-    width: 25px;
-    height: 25px;
+    height: 30px;
+    width: 30px;
+    border-radius: 50%;
     background-color: #FFD400;
-    border-radius: 50%;
-    cursor: pointer;
-    position: relative;
-}
-
-.slider::-moz-range-thumb {
-    -moz-appearance: none;
-    appearance: none;
-    width: 25px;
-    height: 25px;
-    background: #FFD400;
-    border-radius: 50%;
+    border: 2px solid #E9F1F7;
     cursor: pointer;
 }
