+++ index.html
@@ -1,23 +1,36 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Password Generator</title>
-    <link rel="stylesheet" href="css/styles.css">
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Inter:ital,opsz@0,14..32;1,14..32&family=Karla:ital,wght@0,800;1,800&display=swap" rel="stylesheet">
+
+    <!-- Meta tags -->
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+
+    <!-- Google Fonts -->
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link href="https://fonts.googleapis.com/css2?family=Inter:ital,opsz@0,14..32;1,14..32&family=Karla:ital,wght@0,800;1,800&display=swap" rel="stylesheet" />
+
+    <!-- Custom CSS -->
+    <link rel="stylesheet" href="css/styles.css" />
 </head>
 <body>
     <div class="container">
         <div class="txt-container">
-            <h2>Generate a <br/> <span>Random Password</span></h2>
+            <h2>
+                Generate a <br />
+                <span>Random Password</span>
+            </h2>
             <p>Never use an insecure password again.</p>
             <button class="btn" id="generate">Generate passwords</button>
             <div class="slider-container">
-                <input type="range" min="5" max="15" value="5" class="slider" id="myRange">
-                <span class="slider-value" id="sliderValue">50</span>
+                <div class="slider-value-container">
+                    <span class="slider-value">8</span>
+                </div>
+                <div class="slider-field">
+                    <input class="slider" type="range" min="8" max="18" value="8" step="1">
+                </div>
             </div>
         </div>
         <div class="password-container">
@@ -25,6 +38,8 @@
             <p class="password-box" id="password-two"></p>
         </div>
     </div>
-    <script src="/script/script.js" defer></script>
+
+    <!-- Custom JavaScript -->
+    <script src="./script/script.js" type="module"></script>
 </body>
 </html>
